+++ tests/test_client_ws_functional.py
@@ -1,13 +1,9 @@
 import asyncio
-import sys
 
 import pytest
 
 import aiohttp
 from aiohttp import hdrs, helpers, web
-
-
-PY_35 = sys.version_info >= (3, 5)
 
 
 @asyncio.coroutine
@@ -278,42 +274,6 @@ def test_close_from_server(loop, test_client):
     assert msg.type == aiohttp.WSMsgType.CLOSED
 
     yield from closed
-
-
-@pytest.mark.skipif(not PY_35, reason="Requires Python 3.5")
-async def test_closed_async_for(loop, test_client):
-
-    closed = helpers.create_future(loop)
-
-    async def handler(request):
-        ws = web.WebSocketResponse()
-        await ws.prepare(request)
-
-        try:
-            ws.send_bytes(b'started')
-            await ws.receive_bytes()
-        finally:
-            closed.set_result(1)
-        return ws
-
-    app = web.Application(loop=loop)
-    app.router.add_route('GET', '/', handler)
-    client = await test_client(app)
-    resp = await client.ws_connect('/')
-
-    messages = []
-    async for msg in resp:
-        messages.append(msg)
-        if b'started' == msg.data:
-            resp.send_bytes(b'ask')
-            await resp.close()
-
-    assert 1 == len(messages)
-    assert messages[0].type == aiohttp.WSMsgType.BINARY
-    assert messages[0].data == b'started'
-    assert resp.closed
-
-    await closed
 
 
 @asyncio.coroutine
