+++ tests/test_py35/test_client_websocket_35.py
@@ -1,7 +1,7 @@
 import pytest
 
 import aiohttp
-from aiohttp import web
+from aiohttp import helpers, web
 
 async def test_client_ws_async_for(loop, test_client):
     items = ['q1', 'q2', 'q3']
@@ -74,3 +74,38 @@ async def test_client_ws_async_with_shortcut(loop, test_server):
             assert msg.data == 'request/answer'
 
         assert ws.closed
+
+
+async def test_closed_async_for(loop, test_client):
+
+    closed = helpers.create_future(loop)
+
+    async def handler(request):
+        ws = web.WebSocketResponse()
+        await ws.prepare(request)
+
+        try:
+            ws.send_bytes(b'started')
+            await ws.receive_bytes()
+        finally:
+            closed.set_result(1)
+        return ws
+
+    app = web.Application(loop=loop)
+    app.router.add_route('GET', '/', handler)
+    client = await test_client(app)
+    resp = await client.ws_connect('/')
+
+    messages = []
+    async for msg in resp:
+        messages.append(msg)
+        if b'started' == msg.data:
+            resp.send_bytes(b'ask')
+            await resp.close()
+
+    assert 1 == len(messages)
+    assert messages[0].type == aiohttp.WSMsgType.BINARY
+    assert messages[0].data == b'started'
+    assert resp.closed
+
+    await closed
